requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.3.1
python-decouple==3.8
numpy==1.26.4
matplotlib==3.8.4
psycopg2-binary==2.9.9
requests==2.31.0
whitenoise==6.6.0
dj-database-url==2.1.0
gunicorn==21.2.0
